Made User Settings its own JS file
Code for user settings was being used via duplicated code in both page.tsx and header.tsx.
Made a separate file for the user settings that can be used for future implementation.
Note* It appears that while the code was duplicated to page.tsx I did not see any evidence that it was actually being used.