'Hide' the underscores that stretch the User Profile.
Currently there is a row of underscores used in the User profile. These are used because the popover doesn't recognize the size of the dropdown, and as such would normally not scale correctly. The underscores are jarring to look at so I put in a quick hack to hide them for now. In the future I want to remove the underscores entirely, but this hack will let us look more professional for the expo.