Firebase implementation and fixed selection bug
Worked on implementing Firebase implementation (non functional). While I think the code should work when connected to fire base, I can not test this due to an error saying FirebaseError: Failed to get document because the client is offline.

Pushing my changes because I fixed a bug with the selection dropdown, previously it was using preview text as the default instead of actually selecting a default value.